+++ packages/design-system/src/components/Accordion/Primitive/types.ts
@@ -3,6 +3,6 @@ import { ButtonIconType } from 'src/components/ButtonIcon/variations/ButtonIcon'
 export type PanelHeaderAction = ButtonIconType<any> & {
 	icon: string;
 	tooltip: string;
-	dataFeature: string;
+	dataFeature?: string;
 	callback: () => unknown;
 };